Refactor: placing everything into proper modules
Fix: the issue where the previous version was relying on the `certify`
module to already be a part of the `boost` system installation.  Now
`certify` will be a proper git submodule.  CMake will assure that it is
always updated. i.e. project configure triggers git to do a submodule
update.

Fix: how all of the sources were just sitting in the root directory. Now
there is a `lib/` and an `app/` directory.  Each of these projects both have
a `src/` and `include/` directory along with `lib/CmakeLists.txt` and
`app/CMakeLists.txt`.  There is also a `/cmake/common.cmake` file which
contains all of the common build settings.