Adding template support and remove view component
Add: input template support.  It is barely working right now so it needs
some help major re-working.  I took way too long to finish this!

Refactor: decided the view component was unnecessary and thus removed it
from the project.  There is still a good deal of independency between
the components but I think 'common' needs to be renamed and if there is
an 'external' api it is non-existent at the moment.

Refactor: much of the code now resides in 'common'.  Again, I'm not
pleased with this and will be moving it to a more appropriate location
in the near future.