Continuing to fix the build on Linux
Remove: Replaced google-test in favor of Catch2. Found that it was always using `g++` in Linux and creating a
        static library. This was creating linking errors when the compiler was set to `clang++`. I'm sure this
        could have been solved given enough research and time, but I did not feel like wasting any more of either.
        Switching to the amalgamated (single) header version of Catch2 solved the issue.

Remove: As a result of removing google-test there is no need for the `cmake/download-project.cmake`, which was
        downloading the google-test project and building it as part of the cmake configure step.

Remove: Removed the reference to the eric niebler's `range-v3` library. It is facinating and I'm looking forward
        to using it as part of a completed C++20 standard template library.  But experimenting with it at this
        point of the project caused more headaches than I'm able to deal with at the moment.

Added:  Added all the various add include and add flag macros/functions into a seperate cmake source file. Using
        this mechanism makes it so there isn't duplicate flags specified and duplicate include paths specified.
        The include paths was the primary problem which needed solving. On the `g++` build in Linux the include
        path for `boost` was the same as the system C++ include path. For some reason this was causing all kinds
        of errors.

Added:  Added the SG_REQUIRE_NOEXCEPT_IN_CPP17 preprocessor flag. This enforces that all scope guards need to be
        implemented with `noexcept`.

Fixed:  Fixed a `strncpy` usage that could have potentially ended up leaving a string without a null terminating
        character. The `g++` compiler warning `-Wstringop-truncation` lit up with the problem once I finally
        specified a release build.