Adding support for DBC (design-by-contract)
Add: This feature should have been added months ago.  The basic
`PRECONDITION` and `CHECK` macros, which enforces programmers API
assumptions, is critical.  This is one of the major reasons that the C++
Standards Committee is doing their best to add contructs (along with
specialized language support) for the DBC.

Add: A custom C++ exception heirarchy with the very base `std::exceptions`
so it would possible to hook into the construction of the exception and
ensure the message contained a stack trace.

Add: The Google `Backward` library to the project for stack trace
support.  I did try to use `boost::backtrace` but it wanted to link
against a number of things.  Now that I've jumped through all the hoops
involved with `Backward` I should revist the `boost::backtrace` library.

Add: The last thing added this week was the `boost::throw_exception`
macro.  I've modified it quite a bit since I needed to add a stack trace
to the exception additional information.

NOTE: The throw_exception code is the only piece which I did not add
unit tests for.  I really need to add test which prove and show how to
catch one of these exceptions and how to pull the information they
contain.