Add correct time zone support to date/time
Add: Added Howard Hinnant's excellent Date library as an external
dependency.  I believe the entire library was accepted into the C++20
standard so once C++20 is widely supported this can be removed.

Fix: The reason for adding the Date library was that as soon as daylight
saving time kicked in this year all my time output was off by an hour
even after passing it through localtime.  Instead of jumping through
ridiculous hoops to fix the problem the Date library solved the issue
in less than 3 lines of code.  Nothing is better than a well written
API and library.