Enforcing compile time check on calls to fmt::format
Refactor: adding `FMT_STRING("...") to all calls of `fmt::format` which
enables compile time checking of the format string and the arguments.
Luckily, since there were hundreds of calls to `fmt::format`, a single
`sed` search and replace made quick work of 90-95% of the calls.

Add: adding the `FMT_ENFORCE_COMPILE_STRING` define to each project
which forces an error if a developer adds a call to `fmt::format`
without also wrapping the format string within `FMT_STRING("...")`.

For example:

  previously the usage of format was:
  ```cpp
  PLOG_ERROR << fmt::format("ERROR: unable to find file {}", file_name);
  ```
  now this is changed to:
  ```cpp
  PLOG_ERROR << fmt::format(FMT_STRING("ERROR: unable to find file {}"), file_name);
  ```
  In addition, if any developer adds a `fmt::format` without using the
  FMT_STRING the compiler will fail in a spectacular fashion.