Updating logging component to use source_location
C++20 includes the `std::source_location` component which has a static
`current` method to return the current `file`, `function` and `line`.
This has not been fully implemented for GCC/clang as of yet but there
are function intrisics added to query these attributes.

This change is mirroring the `std::source_location` proposal and adding
code which will enable switching over to the `std` version when it is
ready.