Add default output template when none is specified
Refactored quite a bit of the initial `common/results/` project.  Most
of the changes were made to simplify the code, making it easier to read
and harder to mis-use the class API's.

While there I also added a new test suite to test the output of the
`built-in` default template (i.e. what a user will be given if they do
not pass the `--template <path>` on the command line.

I found, yet another, bug in the logging component.  Yet again, the
compiler did not like the few instances whre a `log` statement did not
include arguments.  For example:

```cpp
LOG_ERROR("this should never happen");                            // compiler fails on this
LOG_ERROR("object '{}' should never cause this to happen", obj);  // log function passes without an error
```

Adding a small overload specialization for the `Log` method fixes the
issue.