Finishing uniform exception handling iniside catch
Building on the last pull request where the throw exception code was
cleaned and refactored to have only the information needed and to throw
that information inside a `std::exception` that inherited from a
`boost::exception`. This enabled the code added during these changes to
use the information stored in the exception (either via `std::what` or
via the `boost::exception` framework) to form a uniform error message.