Simplifing output code by removing libfort
Remove: the libfort 3rd party dependency. I was going to add a
dependency on p-ranav/tabulate since it appeared to be an even more
simple library than libfort. It turned out that when trying to bend my
output to fit into tabulate it ended up being considerably more simple
to just use `std::cout` and tools to codify a solution.

Refactor: the 'common/results' library had, from the beginning,
been much more convoluted than I intended. I already had a bug where the
application "core dumped" when more than 1 active network controller was
added to the system. Instead of tracking this problem down I knew I
needed to simplify my output code to handle all sorts of different
input.

I simply started one day and in a little over 500 lines (much more than
it probably needs to be) and cleaner than what it was in
'common/results' I rewrote the output code. Not that it matters, but the
code is completely backwards compatible and takes the 'input json template'
and uses it to format the discovered computer properties out to a
stream.