Refactor output formatting code (#98)
Remove: the libfort 3rd party dependency. I was going to add a
dependency on p-ranav/tabulate since it appeared to be an even more
simple library than libfort. It turned out that when trying to bend my
output to fit into tabulate it ended up being considerably more simple
to just use std::cout and tools to codify a solution.

Refactor: the common/results library had, from the beginning,
been much more convoluted than I intended. I already had a bug where the
application core dumped when more than 1 active network controller was
added to the system. Instead of tracking this problem down I knew I
needed to simplify my output code to handle all sorts of different
input.

In a few days and in a little over 500 lines (much more than
it probably needs to be) I created a new output mechanism.
This was much more concise than what it was in
common/results. It takes the input json template and uses it to format
the discovered computer properties out to an output stream.

Fixing #88
Fixing #86
Fixing #81